completed all the Bacon stuff
computeBaconNumber and computeBaconPath are now implemented and in order (with some testing done even!)

also, i made sure that Kevin Bacon is added to the db upon establishing connection for the first time in the Neo4jDriverSession.java file.

i.e., if Kevin is not in the db, add him there w/ a fixed id as said in the handout.
if he already is, ignore the above and proceed straight to returning the database driver reference.